Footer, Head, Nav, and test
The test is just an example of how to implement the .inc files. The head is missing the <head></head> tags because the <title> may change and so didnt want to lock in one title. The navigation has no links as of yet and needs to be updated when we have all the page names ect.